+++ packages/content/src/lib/markdown-content-renderer.service.spec.ts
@@ -20,10 +20,10 @@ describe('MarkdownContentRendererService', () => {
   it('should transform markdown and return a TOC', async () => {
     const { service } = setup();
     const content = `
-# Level 1 
+# Level 1
 ## Level 2
 
-lorem ipsum .... 
+lorem ipsum ....
 
 # Level 1
 ## Level 2
@@ -57,42 +57,6 @@ Lorem ipsum 2....
     );
 
     testCode = "```typescript\nconsole.log('Hello, world!');\n```";
-    result = await service.render(testCode);
-
-    expect(result).toContain(
-      '<pre class="language-typescript"><code class="language-typescript">'
-    );
-  });
-
-  it('render should correctly highlight diff code blocks', async () => {
-    const { service } = setup();
-    window.Prism.languages['diff'] = {};
-    let testCode = "```diff-javascript\nconsole.log('Hello, world!');\n```";
-    let result = await service.render(testCode);
-
-    expect(result).toContain(
-      '<pre class="language-diff-javascript diff-highlight"><code class="language-diff-javascript diff-highlight">'
-    );
-
-    testCode = "```diff-typescript\nconsole.log('Hello, world!');\n```";
-    result = await service.render(testCode);
-
-    expect(result).toContain(
-      '<pre class="language-diff-typescript diff-highlight"><code class="language-diff-typescript diff-highlight">'
-    );
-  });
-
-  it('render should fall back to language-only highlighting if `diff` plugin is not imported', async () => {
-    const { service } = setup();
-    window.Prism.languages['diff'] = undefined;
-    let testCode = "```diff-javascript\nconsole.log('Hello, world!');\n```";
-    let result = await service.render(testCode);
-
-    expect(result).toContain(
-      '<pre class="language-javascript"><code class="language-javascript">'
-    );
-
-    testCode = "```diff-typescript\nconsole.log('Hello, world!');\n```";
     result = await service.render(testCode);
 
     expect(result).toContain(